Seperated menus from help command and fixed some bugs
